more dependency checking stuff
we still havent lexed everything and the module system isn't recursive yet but there are some basic checks in place to parse the main module as well as see if all the modules loaded in the main module exist. this needs to be done recursively so the entire dependency tree is scanned.