really basic c-style string support + procedure calls.
this is all hacked in to make printf somewhat work but... it works! the name mangler
is hard coded to not mangle a call to the "printf" function. however name mangling is
basically non existant at this point so i need to implement a proper system to handle this
so that the attribute "no_mangle" works on some functions too.