fixed defers; closes #87
basically we would run the defers after the current symbol table had
been left so we would run a defer which would lookup some type
information and this would fail since we had popped the sym table
containing all the type information.

the fix was to add a post visit (as well as the existing pre visit) to
the visitors blocks. and then we run the defer in the post visit before
we leave the symbol table