completely hacked in function calls for functions in other methods.
NOTE this is a massive hack. so nasty that i will have to be doing some relatively large refactorings to the codebase, including a completely new module system.
the module system will need to be reimplemented due to a lack of information that is required, and general spaghetti-ness. the module system is nearly identical to go, but the way it was implemented is kind of hacky

nonetheless... it works which is cool!